Corrected some warning messages.
H3DCommonFunctions.cmake - Added findH3DProjectsIncludeDirsAndLibraries which can be used to find include and library dir for H3D projects and are meant to replace the setup which checks H3D_USE_DEPENDENCIES_ONLY to know if a target should be used or a find module.
Also added function populateProjectIncludeDirectoriesCacheVar which can be used to setup an internal cache variable that other projects can use to find include directories of a certain target in a metarepos build.
If cmake version is above 2.8.7 the variable set by populateProjectIncludeDirectoriesCacheVar is not used but the target property INCLUDE_DIRECTORIES is used instead.

git-svn-id: https://h3dapi.org:8090/CMake/trunk/CMake@100 deaff870-4328-4c97-957a-1a6c77341feb