FindHAPI.cmake - Update to actually find OpenHapticsRenderer/Chai3DRenderer include dirs based on the HAPI_INCLUDE_DIR cache variable.
H3DCommonFunctions.cmake - Updates to findIncludeDirsAndLibrariesForH3DProjects. First I added REQUIRED_PROJECTS_COMPONENTS argument which can be used to require certain COMPONENTS of an H3D library if find_package is used. And also to not remove duplicates in the libraries output because it will remove debug/release markers for libraries.

T921

git-svn-id: https://h3dapi.org:8090/CMake/trunk/CMake@106 deaff870-4328-4c97-957a-1a6c77341feb