Moved getMSVCPostFix, getDefaultH3DOutputDirectoryName, enableCpp11, handleRenamingVariablesBackwardCompatibility and getH3DLibraryNameToSearchFor to new file H3DUtilityFunctions.cmake.
Renamed H3DExternalSearchPath.cmake to H3DCommonFindModuleFunctions.cmake
Updated all find modules to include the ones needed to use the functions.
I did not add the H3DUtilityFunctions.cmake into H3DCommonFindModuleFunctions.cmake because it was not using any of the files. This will break any find module that uses any of the two functions that moved from H3DCommonFindModuleFunctions.cmake to H3DUtilityFunctions.cmake. Since both those functions are relatively new and have never been in an actual release this is acceptable and I will just update all the other that needs it.

The reason for doing this is that I will want to use some of the functions from H3DCommonFindModuleFunctions.cmake in H3DCommonFunctions.cmake to change name of cache variables.

T845

git-svn-id: https://h3dapi.org:8090/CMake/trunk/CMake@110 deaff870-4328-4c97-957a-1a6c77341feb